# optimize the bootstrap function.