@secastel 2018-01-18
Sorry for the delay! Had to find the time to do internal testing and make sure the results are the same. Everything looks good, thanks so much for the fixes!